Javascript Version
This is the first working version of the projekt. Recently there is no login or auth.
The game starts when you add bet to the input and press the Bet button.
You can see your cards and one of the dealers card. You have 2 option, hit or stand. If you stand the dealer try to win against you.